Conditional user avatar delete if it exsits. Set basic setup for class unit tests

+++ src/db/schemas/user/schema.js
@@ -95,8 +95,10 @@ userSchema.pre('remove', async function() {
 
         const roles = await user.getUserRoles();
 
-        await storageApi.deleteBucketObjects(bucketNames[names.user.modelName], [user.avatar.keyname]);
-
+        if (user.avatar) {
+            await storageApi.deleteBucketObjects(bucketNames[names.user.modelName], [user.avatar.keyname]);
+        }
+        
         await user.model(names.userRole.modelName).deleteMany({
             user: user._id
         });
